+++ spec/Gaufrette/FilesystemSpec.php
@@ -204,7 +204,7 @@ class FilesystemSpec extends ObjectBehavior
     function it_fails_when_delete_is_not_successful(Adapter $adapter)
     {
         $adapter->exists('filename')->willReturn(true);
-        $adapter->delete('filename')->willThrow(StorageFailure::class);
+        $adapter->delete('filename')->willThrow(StorageFailure::unexpectedFailure('delete', []));
 
         $this
             ->shouldThrow(StorageFailure::class)